feat(benchmark): 重构性能基准测试系统并添加实时性能报告
  - 重构核心性能追踪类型,统一使用独立的周期ID
  - 新增完整的节点和策略性能报告体系,支持详细的统计分析
  - 实现基于SSE的实时性能数据推送
  - 为所有策略节点添加完整的性能追踪支持
  - 优化事件系统,移除不必要的反序列化trait
  - 修复变量节点性能追踪缺失的问题

+++ engine/src/backtest_strategy_engine/node/if_else_node/if_else_node_context/mod.rs
@@ -92,7 +92,7 @@ impl IfElseNodeContext {
 
     // 开始评估各个分支
     pub async fn evaluate(&mut self) -> Result<(), IfElseNodeError> {
-        let mut cycle_tracker = CycleTracker::new(self.get_play_index());
+        let mut cycle_tracker = CycleTracker::new(self.get_play_index() as u32);
 
 
         let mut case_matched = false; // 是否匹配到case
